Another alternative to solving the time problem
