+++ app/src/main/java/com/example/wallet/di/HomeModule.kt
@@ -6,7 +6,7 @@ import com.example.wallet.feature_add_expense.domain.usecase.AddExpenseUseCase
 import com.example.wallet.feature_add_expense.domain.usecase.AddUseCases
 import com.example.wallet.feature_home.data.repository.HomeExpenseRepositoryImpl
 import com.example.wallet.feature_home.domain.repository.HomeExpenseRepository
-import com.example.wallet.feature_main.domain.time.TimeManager
+import com.example.wallet.feature_main.domain.time.TimeProvider
 import dagger.Module
 import dagger.Provides
 import dagger.hilt.InstallIn
@@ -24,8 +24,8 @@ object HomeModule {
     @Provides
     fun providesAddUseCase(
         repository: CreateExpenseRepository,
-        timeManager: TimeManager
-    ) = AddExpenseUseCase(repository, timeManager)
+        timeProvider: TimeProvider
+    ) = AddExpenseUseCase(repository, timeProvider)
 
     @Provides
     fun providesAddUseCases(addUseCase: AddExpenseUseCase) = AddUseCases(addUseCase)
