+++ app/src/main/java/com/example/wallet/feature_main/domain/time/TimeProvider.kt
@@ -2,7 +2,7 @@ package com.example.wallet.feature_main.domain.time
 
 import java.time.Clock
 
-interface TimeManager {
+interface TimeProvider {
     fun now(): Time
     fun now(millis: Long): Time
     fun now(clock: Clock): Time
